requirements.txt
streamlit>=1.36.0
google-genai>=1.0.0
groq>=0.5.0
google-api-python-client>=2.100.0
requests>=2.31.0
pandas>=2.0.0
